Implement finer grained handle locking
Create a RwHashes type that abstracts away creating a large number
of HashMap's in parallel and using one of them depending on the
hash of the key being used. That way we don't get a bunch of lock
contention when multiple threads are all trying to read() or write()
and thus all want to grab a handle.