Make sure we mark blobs as local on prefetch
To make sure anything that's prefetched gets marked as local we
make sure that when blobs are touched that happens. The counting
isn't as exact but the errors are all on the side of marking local
something that isn't yet and the removals path will just try and
remove a non-existant file which is fine.