Make sure touch() creates blobs in the DB
When we're fetching blobs that didn't exist in the DB yet make
touch() create them. That way we have a single path that does this
and are not only resilient to crashes but enable an easy way to do
cloning of repositories and syncing. We no longer need to have
anything in the blobs table for things to work.